+++ lib/util/mixins/move_to_position_along_the_path.dart
@@ -8,6 +8,8 @@ import 'package:flutter/material.dart';
 
 /// Mixin responsible for find path using `a_star_algorithm` and moving the component through the path
 mixin MoveToPositionAlongThePath on Movement {
+  static const REDUCTION_SPEED_DIAGONAL = 0.7;
+
   List<Offset> _currentPath = [];
   int _currentIndex = 0;
   bool _showBarriers = false;
@@ -93,14 +95,30 @@ mixin MoveToPositionAlongThePath on Movement {
         stopMoveAlongThePath();
       }
     } else {
-      if (diffX > 0 && diffX.abs() > 0.5) {
-        this.moveRight(displacementX);
-      } else if (diffX < 0 && diffX.abs() > 0.5) {
-        this.moveLeft(displacementX);
-      } else if (diffY > 0 && diffY.abs() > 0.5) {
-        this.moveDown(displacementY);
-      } else if (diffY < 0 && diffY.abs() > 0.5) {
-        this.moveUp(displacementY);
+      if (diffX.abs() > 0.5 && diffY.abs() > 0.5) {
+        final displacementXDiagonal = displacementX * REDUCTION_SPEED_DIAGONAL;
+        final displacementYDiagonal = displacementY * REDUCTION_SPEED_DIAGONAL;
+        if (diffX > 0 && diffY > 0) {
+          this.moveDownRight(displacementXDiagonal, displacementYDiagonal);
+        } else if (diffX < 0 && diffY > 0) {
+          this.moveDownLeft(displacementXDiagonal, displacementYDiagonal);
+        } else if (diffX > 0 && diffY < 0) {
+          this.moveUpRight(displacementXDiagonal, displacementYDiagonal);
+        } else if (diffX < 0 && diffY < 0) {
+          this.moveUpLeft(displacementXDiagonal, displacementYDiagonal);
+        }
+      } else if (diffX.abs() > 0.5) {
+        if (diffX > 0) {
+          this.moveRight(displacementX);
+        } else if (diffX < 0) {
+          this.moveLeft(displacementX);
+        }
+      } else if (diffY.abs() > 0.5) {
+        if (diffY > 0) {
+          this.moveDown(displacementY);
+        } else if (diffY < 0) {
+          this.moveUp(displacementY);
+        }
       }
     }
   }